+++ src/data/tracking.ts
@@ -13,4 +13,5 @@ export const buttonLocation = {
     bodyDemo: 3,
     trySourcegraph: 4,
     body: 5,
+    caseStudy: 6,
 }
